+++ packages/ice/src/config.ts
@@ -1,11 +1,11 @@
 import { createRequire } from 'module';
 import trustCert from '@ice/bundles/compiled/trusted-cert/index.js';
 import fse from 'fs-extra';
-import consola from 'consola';
 import type { Config } from '@ice/webpack-config/esm/types';
 import type { UserConfigContext } from 'build-scripts';
 import lodash from '@ice/bundles/compiled/lodash/index.js';
 import type { UserConfig } from './types/userConfig.js';
+import { logger } from './utils/logger.js';
 
 const require = createRequire(import.meta.url);
 const { merge } = lodash;
@@ -122,7 +122,7 @@ const userConfig = [
   {
     name: 'ssr',
     validation: 'boolean',
-    defaultValue: true,
+    defaultValue: false,
   },
   {
     name: 'server',
@@ -133,7 +133,7 @@ const userConfig = [
     },
     setConfig: (_config: Config, server: UserConfig['server']) => {
       if (server.format === 'esm' && server.bundle) {
-        consola.error('Do not support bundle in ESM mode. Please set `server.bundle` to false.');
+        logger.error('Do not support bundle in ESM mode. Please set `server.bundle` to false.');
         process.exit(1);
       }
     },
@@ -153,7 +153,7 @@ const userConfig = [
       if (configureWebpack) {
         // create warning for user
         const customConfigWebpack: Config['configureWebpack'][0] = (...args) => {
-          consola.warn('It is not recommended to configure webpack directly.');
+          logger.warn('It is not recommended to configure webpack directly.');
           return configureWebpack(...args);
         };
         config.configureWebpack = [...(config.configureWebpack || []), customConfigWebpack];
@@ -197,7 +197,7 @@ const userConfig = [
           },
         });
       } else {
-        consola.warn(`dropLogLevel only support [${Object.keys(levels).join(',')}]`);
+        logger.warn(`dropLogLevel only support [${Object.keys(levels).join(',')}]`);
       }
     },
   },
@@ -271,7 +271,7 @@ const userConfig = [
         const dependenciesMsg = 'Please check dependencies of eslint(> 7.0.0)';
 
         if (dependencyError) {
-          consola.warn(dependenciesMsg);
+          logger.warn(dependenciesMsg);
           return;
         }
         let eslintOptions = {
@@ -358,7 +358,7 @@ const userConfig = [
 
 const cliOption = [
   {
-    name: 'platform',
+    name: 'target',
     commands: ['start', 'build'],
   },
   {