+++ packages/ice/src/service/serverCompiler.ts
@@ -1,5 +1,4 @@
 import * as path from 'path';
-import consola from 'consola';
 import { esbuild } from '@ice/bundles';
 import fse from 'fs-extra';
 import fg from 'fast-glob';
@@ -22,9 +21,13 @@ import isExternalBuiltinDep from '../utils/isExternalBuiltinDep.js';
 import getServerEntry from '../utils/getServerEntry.js';
 import type { DepScanData } from '../esbuild/scan.js';
 import formatPath from '../utils/formatPath.js';
+import { createLogger } from '../utils/logger.js';
+import { getExpandedEnvs } from '../utils/runtimeEnv.js';
 import { scanImports } from './analyze.js';
 import type { DepsMetaData } from './preBundleCJSDeps.js';
 import preBundleCJSDeps from './preBundleCJSDeps.js';
+
+const logger = createLogger('server-compiler');
 
 interface Options {
   rootDir: string;
@@ -66,10 +69,12 @@ export function createServerCompiler(options: Options) {
     preBundle,
     swc,
     externalDependencies,
-    transformEnv = true,
     compilationInfo,
     redirectImports,
     removeOutputs,
+    runtimeDefineVars = {},
+    enableEnv = false,
+    transformEnv = true,
   } = {}) => {
     let depsMetadata: DepsMetaData;
     let swcOptions = merge({}, {
@@ -84,7 +89,7 @@ export function createServerCompiler(options: Options) {
     const transformPlugins = getCompilerPlugins({
       ...task.config,
       fastRefresh: false,
-      env: false,
+      enableEnv,
       polyfill: false,
       swcOptions,
       redirectImports,
@@ -106,7 +111,6 @@ export function createServerCompiler(options: Options) {
     }
 
     // get runtime variable for server build
-    const runtimeDefineVars = {};
     Object.keys(process.env).forEach((key) => {
       // Do not transform env when bundle client side code.
       if (/^ICE_CORE_/i.test(key) && transformEnv) {
@@ -120,6 +124,18 @@ export function createServerCompiler(options: Options) {
       ...defineVars,
       ...runtimeDefineVars,
     };
+    const expandedEnvs = getExpandedEnvs();
+    // Add user defined envs.
+    for (const [key, value] of Object.entries(expandedEnvs)) {
+      define[`import.meta.env.${key}`] = JSON.stringify(value);
+    }
+    // Add process.env.
+    Object.keys(process.env)
+      .filter((key) => /^ICE_/.test(key) || key === 'NODE_ENV')
+      .forEach((key) => {
+        define[`import.meta.env.${key}`] = JSON.stringify(process.env[key]);
+      });
+
     const format = customBuildOptions?.format || 'esm';
 
     let buildOptions: esbuild.BuildOptions = {
@@ -138,6 +154,12 @@ export function createServerCompiler(options: Options) {
       define,
       absWorkingDir: rootDir,
       external: Object.keys(externals),
+      banner: customBuildOptions.platform === 'node' && server?.format !== 'cjs'
+        ? {
+            // See https://github.com/evanw/esbuild/issues/1921#issuecomment-1152991694
+            js: 'import { createRequire } from \'module\';const require = createRequire(import.meta.url);',
+          }
+        : undefined,
       plugins: [
         ...(customBuildOptions.plugins || []),
         emptyCSSPlugin(),
@@ -173,12 +195,12 @@ export function createServerCompiler(options: Options) {
     }
 
     const startTime = new Date().getTime();
-    consola.debug('[esbuild]', `start compile for: ${JSON.stringify(buildOptions.entryPoints)}`);
+    logger.debug('[esbuild]', `start compile for: ${JSON.stringify(buildOptions.entryPoints)}`);
 
     try {
       const esbuildResult = await esbuild.build(buildOptions);
 
-      consola.debug('[esbuild]', `time cost: ${new Date().getTime() - startTime}ms`);
+      logger.debug('[esbuild]', `time cost: ${new Date().getTime() - startTime}ms`);
 
       const esm = server?.format === 'esm';
       const outJSExtension = esm ? '.mjs' : '.cjs';
@@ -198,9 +220,14 @@ export function createServerCompiler(options: Options) {
         serverEntry,
       };
     } catch (error) {
-      consola.error('Server compile error.', `\nEntryPoints: ${JSON.stringify(buildOptions.entryPoints)}`);
-      consola.debug('Build options: ', buildOptions);
-      consola.debug(error.stack);
+      logger.error(
+        'Server compile error.',
+        `\nEntryPoints: ${JSON.stringify(buildOptions.entryPoints)}`,
+        `\n${error.message}`,
+      );
+      // TODO: Log esbuild options with namespace.
+      // logger.debug('esbuild options: ', buildOptions);
+      logger.debug(error.stack);
       return {
         error: error as Error,
       };