+++ packages/ice/src/types/plugin.ts
@@ -48,6 +48,8 @@ export type ServerCompiler = (
     };
     redirectImports?: Config['redirectImports'];
     removeOutputs?: boolean;
+    runtimeDefineVars?: Record<string, string>;
+    enableEnv?: boolean;
   }
 ) => Promise<ServerBuildResult>;
 export type WatchEvent = [