+++ packages/ice/src/utils/getServerCompilerPlugin.ts
@@ -11,12 +11,25 @@ interface Options {
   userConfig: UserConfig;
   outputDir: string;
   serverEntry: string;
+  incremental: boolean;
   dataCache: Map<string, string>;
   serverCompileTask: ExtendsPluginAPI['serverCompileTask'];
   ensureRoutesConfig: () => Promise<void>;
+  runtimeDefineVars: Record<string, string>;
 }
+
 function getServerCompilerPlugin(serverCompiler: ServerCompiler, options: Options) {
-  const { outputDir, rootDir, serverEntry, userConfig, dataCache, serverCompileTask, ensureRoutesConfig } = options;
+  const {
+    outputDir,
+    rootDir,
+    serverEntry,
+    userConfig,
+    dataCache,
+    serverCompileTask,
+    ensureRoutesConfig,
+    runtimeDefineVars,
+    incremental,
+  } = options;
   const entryPoint = getServerEntry(rootDir, serverEntry);
   const { ssg, ssr, server: { format } } = userConfig;
   const isEsm = userConfig?.server?.format === 'esm';
@@ -28,14 +41,12 @@ function getServerCompilerPlugin(serverCompiler: ServerCompiler, options: Option
         outdir: path.join(outputDir, SERVER_OUTPUT_DIR),
         splitting: isEsm,
         format,
-        // When format set to esm and platform set to node,
-        // will cause error `Dynamic require of "XXX" is not supported`.
-        // https://github.com/evanw/esbuild/issues/1921
-        platform: isEsm ? 'browser' : 'node',
+        platform: 'node',
         mainFields: ['module', 'main'],
         outExtension: { '.js': isEsm ? '.mjs' : '.cjs' },
         metafile: true,
         logLevel: 'silent', // The server compiler process will log it in debug.
+        incremental,
       },
       {
         preBundle: format === 'esm' && (ssr || ssg),
@@ -47,6 +58,7 @@ function getServerCompilerPlugin(serverCompiler: ServerCompiler, options: Option
           },
         },
         removeOutputs: true,
+        runtimeDefineVars,
       },
     ],
     ensureRoutesConfig,