+++ packages/runtime/src/Document.tsx
@@ -146,7 +146,7 @@ export function Data(props: DataProps) {
   const routePath = getCurrentRoutePath(matches);
   const windowContext: WindowContext = {
     appData,
-    routesData: renderMode === 'SSG' ? null : routesData,
+    routesData,
     routesConfig,
     routePath,
     downgrade,