+++ packages/bundles/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.3
+
+- [fix] Bump version of `@ice/swc-plugin-keep-export`(0.1.3 -> 0.1.4), https://github.com/ice-lab/swc-plugins/pull/8
+
+## 0.1.2
+
+- [fix] add esbuild and esbuild-register
+
 ## 0.1.1
 
 - [fix] Bump version of `@swc/core`(1.3.3 -> 1.3.19), https://github.com/swc-project/swc/issues/6371