+++ packages/runtime/CHANGELOG.md
@@ -4,6 +4,15 @@
 
 - [feat] suspense ssr
 
+## v1.0.5
+
+- [feat] support template parse for dataLoader
+
+## v1.0.4
+
+- [chore] show more log when failed to load route module
+- [fix] rerender after hydrate for ssg page
+
 ## v1.0.3
 
 - [feat] optimize log when load route module failed